+++ app/uk/gov/hmrc/softdrinksindustrylevystub/services/DesSubmissionService.scala
@@ -27,7 +27,7 @@ import scala.collection.mutable
 class DesSubmissionService {
 
   lazy val store: mutable.Map[String, Subscription] = mutable.Map.empty
-  lazy private val returnStore: mutable.Map[String, Return] = mutable.Map.empty
+  lazy val returnStore: mutable.Map[String, Return] = mutable.Map.empty
 
   def createSubscriptionResponse(idNumber: String, data: Subscription): CreateSubscriptionResponse = {
     import uk.gov.hmrc.softdrinksindustrylevystub.models.EnumUtils.idEnum
